Rename component, call api to generate email
