+++ src/components/RequestPasswordReset/RequestPasswordResetStyle.tsx
@@ -1,6 +1,6 @@
 import styled from "styled-components";
 
-import { Button } from "../../components/Button/Button";
+import { Button } from "../Button/Button";
 
 export const StyledFormContainer = styled.div`
   border: ${props => props.theme.border};
